Makr add logging (#362)
* chore: add logging to primer service

* chore: add logging to gpio_mw

* chore: add i2c mw

* chore: add temp logging

* chore: add logging to timestamp

* refactor: some ara::log cleanup

+++ apps/fc/radio_service/radio_app.cc
@@ -149,12 +149,12 @@ void RadioApp::OnRadioStatusMsg(const mavlink_message_t& msg) {
   mavlink_msg_radio_status_decode(&msg, &radio_status);
   ara::log::LogWarn()
       << "rxErrors: "      << radio_status.rxerrors
-      << " TxFreeBuf: "    << static_cast<uint32_t>(radio_status.txbuf)
+      << " TxFreeBuf: "    << radio_status.txbuf
       << " Fixed: "        << radio_status.fixed
-      << " rssi: "         << static_cast<uint32_t>(radio_status.rssi)
-      << " remote rssi: "  << static_cast<uint32_t>(radio_status.remrssi)
-      << " noise: "        << static_cast<uint32_t>(radio_status.noise)
-      << " remote noise: " << static_cast<uint32_t>(radio_status.remnoise);
+      << " rssi: "         << radio_status.rssi
+      << " remote rssi: "  << radio_status.remrssi
+      << " noise: "        << radio_status.noise
+      << " remote noise: " << radio_status.remnoise;
 
   RadioDataType someip_radio_data {
       .rxerrors = radio_status.rxerrors,
